nstall to more local folder
Also add that as a package search path

+++ scripts/zlib_build_qnx.sh
@@ -7,12 +7,17 @@ if [[ -z "${QNX_SOURCE_SCRIPT}" ]]; then
     exit 1
 fi
 
+if [[ -z "${LDK_QNX_INSTALL_FOLDER}" ]]; then
+    echo "Set the environment variable LDK_QNX_INSTALL_FOLDER to specify where dependencies should be installed to."
+    exit 1
+fi
+
 . "${QNX_SOURCE_SCRIPT}"
 
 rm -rf build
 mkdir build
 cd build
 
-cmake -G "Unix Makefiles" .. -DCPPZMQ_BUILD_TESTS=OFF -DCMAKE_BUILD_TYPE=Release -DCMAKE_TOOLCHAIN_FILE=cmake/toolchains/aarch64/qnx/gcc/toolchain_aarch64_qnx_gcc.cmake
+cmake -G "Unix Makefiles" .. -DCMAKE_PREFIX_PATH:PATH=${LDK_QNX_INSTALL_FOLDER} -DCMAKE_INSTALL_PREFIX:PATH=${LDK_QNX_INSTALL_FOLDER} -DCMAKE_BUILD_TYPE=Release -DCMAKE_TOOLCHAIN_FILE=cmake/toolchains/aarch64/qnx/gcc/toolchain_aarch64_qnx_gcc.cmake
 
-cmake --build . -- -j "${LUMPDK_NPROC:-$(($(nproc) + 2))}"
+cmake --build . --target install -- -j "${LUMPDK_NPROC:-$(($(nproc) + 2))}"
